a little bit of work on efficiency
(now we're up to 75MiB/s out of the old 111MiB/s; have run out of easy ideas on what could cause slowdowns)

+++ py/sensornet/test.py
@@ -210,7 +210,7 @@ async def test11():
     h = sn.Handler(8,8,8,8, 64)
     async def data_later():
         await asyncio.sleep(.2)
-        h.set('hey listen', np.zeros(128), sn.Float(128))
+        h.set('hey listen', np.zeros(128), sn.Float(2, 128))
     async def query_later():
         await asyncio.sleep(.2)
         assert (await h.query('forces of evil gather', sn.Float(16))) is None
